requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
slowapi>=0.1.9
pydantic>=2.0.0
lighter-sdk>=1.0.0
